+++ lib/ReactViews/Tools/DiffTool/DiffTool.tsx
@@ -38,7 +38,7 @@ import Button, { RawButton } from "../../../Styled/Button";
 import Select from "../../../Styled/Select";
 import Spacing from "../../../Styled/Spacing";
 import Text, { TextSpan } from "../../../Styled/Text";
-import RasterLayerTraits from "../../../Traits/TraitsClasses/RasterLayerTraits";
+import ImageryProviderTraits from "../../../Traits/TraitsClasses/ImageryProviderTraits";
 import { parseCustomMarkdownToReactWithOptions } from "../../Custom/parseCustomMarkdownToReact";
 import { GLYPHS, StyledIcon } from "../../../Styled/Icon";
 import Loader from "../../Loader";
@@ -1068,8 +1068,8 @@ function setTimeFilterFromLocation(
 
 function hasOpacity(
   model: any
-): model is HasTrait<RasterLayerTraits, "opacity"> {
-  return hasTraits(model, RasterLayerTraits, "opacity");
+): model is HasTrait<ImageryProviderTraits, "opacity"> {
+  return hasTraits(model, ImageryProviderTraits, "opacity");
 }
 
 export default hoistStatics(withTranslation()(withTheme(DiffTool)), DiffTool);