+++ lib/Traits/TraitsClasses/MapboxMapCatalogItemTraits.ts
@@ -3,10 +3,10 @@ import mixTraits from "../mixTraits";
 import CatalogMemberTraits from "./CatalogMemberTraits";
 import LegendOwnerTraits from "./LegendOwnerTraits";
 import MappableTraits from "./MappableTraits";
-import RasterLayerTraits from "./RasterLayerTraits";
+import ImageryProviderTraits from "./ImageryProviderTraits";
 
 export default class MapboxMapCatalogItemTraits extends mixTraits(
-  RasterLayerTraits,
+  ImageryProviderTraits,
   CatalogMemberTraits,
   MappableTraits,
   LegendOwnerTraits
@@ -31,14 +31,6 @@ export default class MapboxMapCatalogItemTraits extends mixTraits(
     type: "string"
   })
   accessToken?: string;
-
-  @primitiveTrait({
-    name: "maximumLevel",
-    description:
-      "The maximum level-of-detail supported by the imagery provider, or undefined if there is no limit.",
-    type: "number"
-  })
-  maximumLevel = 25;
 
   @primitiveTrait({
     name: "format",